Move collation generation code into main overseer module
